requirements.txt
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]==0.89.1
loguru
pytest
pytest-cov
opencv-python
pandas
numpy
Pillow